requirements.txt
python-dotenv
boto3
jwt